+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 1.0.2 (2017-02-13)
+## 1.0.2 (2017-02-15)
 
 * Provide some `Bundler` implementations for lists, see #4
 * Add a workaround for the Jack compiler where the package name is always "package", may fix #5
+* Erase generics of extended classes, fixes #6
 
 ## 1.0.1 (2017-01-27)